Consolidated files, added subrouters and base routes
