+++ pkg/client/create_user_channel.go
@@ -6,25 +6,28 @@ import (
 	"errors"
 	"github.com/automuteus/galactus/pkg/endpoint"
 	"github.com/bwmarrin/discordgo"
+	"go.uber.org/zap"
 	"io/ioutil"
-	"log"
 	"net/http"
 )
 
 func (galactus *GalactusClient) CreateUserChannel(userID string) (*discordgo.Channel, error) {
-	resp, err := galactus.client.Post(galactus.Address+endpoint.UserChannelCreatePartial+userID, "application/json", bytes.NewBufferString(""))
+	url := endpoint.FormGalactusURL(galactus.Address, endpoint.DiscordRoute, endpoint.UserChannelCreatePartial, userID)
+	resp, err := galactus.client.Post(url, "application/json", bytes.NewBufferString(""))
 	if err != nil {
 		return nil, err
 	}
 	respBytes, err := ioutil.ReadAll(resp.Body)
 	if err != nil {
-		log.Println("error reading all bytes from resp body for createUserChannel")
-		log.Println(err)
+		galactus.logger.Error("error reading all bytes from message body",
+			zap.Error(err),
+			zap.String("url", url),
+		)
 		return nil, err
 	}
 	defer resp.Body.Close()
 	if resp.StatusCode != http.StatusOK {
-		err := errors.New("non-200 status code received for createUserChannel:")
+		err := errors.New("non-200 response code received for " + url)
 		return nil, err
 	}
