+++ pkg/endpoint/capture.go
@@ -1,7 +1,9 @@
 package endpoint
 
-const RequestJob = "/request/job"
+const CaptureRoute = "/api/capture"
+
 const JobCount = "/totalJobs"
+const RequestJob = "/request/job"
 
 const AddCaptureEventPartial = "/addCaptureEvent/"
 const AddCaptureEventFull = AddCaptureEventPartial + "{connectCode}/{eventType}"
@@ -14,6 +16,3 @@ const GetCaptureTaskFull = GetCaptureTaskPartial + "{connectCode}"
 
 const SetCaptureTaskStatusPartial = "/setCaptureTaskStatus/"
 const SetCaptureTaskStatusFull = SetCaptureTaskStatusPartial + "{taskID}"
-
-const GetGuildAMUSettingsPartial = "/getAMUSettings/"
-const GetGuildAMUSettingsFull = GetGuildAMUSettingsPartial + "{guildID}"
