+++ pkg/endpoint/general.go
@@ -1,0 +1,3 @@
+package endpoint
+
+const GeneralRoute = "/api"
