+++ pkg/endpoint/settings.go
@@ -1,0 +1,6 @@
+package endpoint
+
+const SettingsRoute = "/api/settings"
+
+const GetGuildAMUSettingsPartial = "/getAMUSettings/"
+const GetGuildAMUSettingsFull = GetGuildAMUSettingsPartial + "{guildID}"
