+++ pkg/endpoint/stats.go
@@ -1,0 +1,3 @@
+package endpoint
+
+const StatsRoute = "/api/stats"
